update på matador

+++ Arbejdsmappe/IntelliJ/bank/src/BankAccount.java
@@ -1,5 +1,5 @@
 public class BankAccount {
-    private float balance = 30000;
+    private int balance = 30000;
     private String owner;
     private int stalkingCounter = 0;
 
@@ -8,7 +8,7 @@ public class BankAccount {
      *
      * @param owner
      */
-    public BankAccount(String owner, float balance) {
+    public BankAccount(String owner, int balance) {
         this.owner = owner;
         this.balance = balance;
     }
@@ -19,7 +19,7 @@ public class BankAccount {
      *
      * @param amount
      */
-    public void doTransaction(float amount) {
+    public void doTransaction(int amount) {
 
         balance += amount; // same as balance = balance+ amount;
     }
@@ -39,7 +39,7 @@ public class BankAccount {
     }
 
 
-    public float getBalance() {
+    public int getBalance() {
 
         stalkingCounter++;// Den her linje skal demonstrere en af de ting jeg som programmør kan få ud af begrænse adgangen klassens felter (bemærk feltet 'balance' i toppen af klassen er private).
         // Hver gang instansens saldo skal 'ses' udefra, skal det ske gennem denne metode og jeg kan så tilføje lidt  kode
